pyproject.toml
[project]
name = "chataui-app"
version = "0.1.0"
description = "A Advance Question Answering Stateful chatbot with GitHub and Google authentication built with Python, UV, and Chainlit."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.8.1",
    "openai-agents>=0.2.0",
    "python-dotenv>=1.0.1",
]
